LayoutTests/imported/w3c/web-platform-tests/accname/name/comp_name_from_content-expected.txt

Tests the #comp_name_from_content portions of the AccName Name Computation algorithm.

This series of tests exercises the button, heading, and link elements, because each have different characteristics worthy of testing in each of the name computation algorithm scenarios:

button is a leaf node with sub-level elements presentational.
heading is block level, and can contain sub-level interactives like links.
link (a[href]) is an interactive inline element that can include non-presentational descendants.
ARIA roles, inline

label
label
label

ARIA roles, block

label
label
label

simple

label
label

label

simple with ::before

label
label

label

simple with ::after

label
label

label

simple with ::before and ::after

label
label

label

no space joiners ::before and ::after

label
label

label

Arabic right-to-left (RTL) with ::before and ::after

اسم
اسم

اسم

Alternative Text for CSS content (previously `alt:`) in pseudo-elements

rendered text should be "before label after"

accessibility label should be "alt-before label alt-after"

label
label

label

Mixed Alternative Text (attr() and strings) for CSS content (previously `alt:`) in pseudo-elements

rendered text should be "before label after"

accessibility label should be "start alt-before end label start alt-after end"

label
label

label

Empty alternative text for CSS content in pseudo-elements when applied to primitive appearance form controls





simple w/ for each child

one two three
one two three

one two three

simple for each child with image

one  three
one  three

one  three

simple for each child with extra nesting containing image

one two  four
one two  four

one two  four

heading with nested button with nested image

heading  button   button  heading

heading with nested link with nested image

heading link   link heading

heading with nested link with nested image using aria-label

heading ignored link text   ignored link text heading

heading with nested link with nested image using aria-labelledby

heading ignored link text   ignored link text heading

heading with two nested links referencing image using aria-labelledby

link1 link2   link3

simple w/ for each child (w/o spaces and display:inline)

onetwothree
onetwothree

onetwothree

simple w/ for each child (w/o spaces and display:block)


one
two
three

one
two
three

one
two
three


simple w/ for each child (w/o spaces and display:inline block)

onetwothree
onetwothree

onetwothree


PASS aria button name from content, inline
PASS aria heading name from content, inline
PASS aria link name from content, inline
PASS aria button name from content, block
PASS aria heading name from content, block
PASS aria link name from content, block
PASS button name from content
PASS heading name from content
PASS link name from content
PASS button name from content with ::before
PASS heading name from content with ::before
PASS link name from content with ::before
PASS button name from content with ::after
PASS heading name from content with ::after
PASS link name from content with ::after
PASS button name from content with ::before and ::after
PASS heading name from content with ::before and ::after
PASS link name from content with ::before and ::after
PASS button name from content no space joiners ::before and ::after
PASS heading name from content no space joiners ::before and ::after
PASS link name from content no space joiners ::before and ::after
PASS button name from content with ::before and ::after in rtl
PASS heading name from content with ::before and ::after in rtl
PASS link name from content with ::before and ::after in rtl
PASS button name from fallback content with ::before and ::after
PASS heading name from fallback content with ::before and ::after
PASS link name from fallback content with ::before and ::after
PASS button name from fallback content mixing attr() and strings with ::before and ::after
PASS heading name from fallback content mixing attr() and strings with ::before and ::after
PASS link name from fallback content mixing attr() and strings with ::before and ::after
PASS primitive radio input with ::before containing empty alternative text
PASS primitive radio input with ::before containing empty alternative text for an image
PASS button name from content for each child
PASS heading name from content for each child
PASS link name from content for each child
PASS button name from content for each child including image
PASS heading name from content for each child including image
PASS link name from content for each child including image
PASS button name from content for each child including nested image
PASS heading name from content for each child including nested image
PASS link name from content for each child including nested image
PASS heading name from content for each child including nested button with nested image
PASS heading name from content for each child including nested link with nested image
PASS heading name from content for each child including nested link using aria-label with nested image
PASS heading name from content for each child including nested link using aria-labelledby with nested image
FAIL heading name from content for each child including two nested links using aria-labelledby with nested image assert_equals: <h3 data-expectedlabel="image link2 link3" data-testname="heading name from content for each child including two nested links using aria-labelledby with nested image" class="ex">
  <a href="#" aria-labelledby="nested_image_label2">
    link1<!-- this text is skipped because of aria-labelledby -->
  </a>
  <a href="#" data-expectedlabel="link2 image link3" data-testname="link name from content for each child including nested image (referenced elsewhere via labelledby)" class="ex">
    link2
    <img id="nested_image_label2" alt="image" src="data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==">
    <!-- image skipped in this link (when computing heading text) because it was already referenced by the first link within this heading label recursion cycle. -->
    <!-- but image not skipped when computing the text of the link itself since it has not been referenced in that context -->
    link3
  </a>
</h3> expected "image link2 link3" but got "image link2 image link3"
PASS link name from content for each child including nested image (referenced elsewhere via labelledby)
PASS button name from content for each child (no space, inline)
PASS heading name from content for each child (no space, inline)
PASS link name from content for each child (no space, inline)
PASS button name from content for each child (no space, display:block)
PASS heading name from content for each child (no space, display:block)
PASS link name from content for each child (no space, display:block)
PASS button name from content for each child (no space, display:inline-block)
PASS heading name from content for each child (no space, display:inline-block)
PASS link name from content for each child (no space, display:inline-block)
